+++ src/index.js
@@ -1,55 +1,16 @@
-import { createBioformatsZarrLoader, createOMETiffLoader } from '@hms-dbmi/viv';
-import { images } from './test_config.json';
-import { getTileIndices } from "./vendored-utils";
-async function timeGetTile(loader, { x, y, z, loaderSelection }) {
-  const start = performance.now()
-  await loader.getTile({ x, y, z, loaderSelection });
-  const end = performance.now();
-  return [start, end];
-}
+import { range, timeGetTile, getTileCoords, getLoader } from './utils';
+import { images } from './config.json';
 
-async function getLoader({ url, format }) {
-  if (format === 'zarr') {
-    return createBioformatsZarrLoader({ url });
-  }
-  return createOMETiffLoader({ url });
-}
+const ITERS = 10;
 
-function getTileCoords(props) {
-  const {
-    xCoord,
-    yCoord,
-    width,
-    height,
-    zoom: originalZoom,
-    tileSize,
-    extent,
-  } = props;
-  const tileIndices = [];
-  for (let zoom = originalZoom; zoom <= 0; zoom++) {
-    const viewState = { target: [xCoord, yCoord], zoom };
-    tileIndices.push(
-      ...getTileIndices({ viewState, width, height, tileSize, extent })
-    );
-  }
-  // See https://github.com/visgl/deck.gl/pull/4616/files#diff-4d6a2e500c0e79e12e562c4f1217dc80R128
-  // and https://github.com/hms-dbmi/viv/blob/db6f08a14f4d49590c49ea6ca23055782dcab29a/src/layers/MultiscaleImageLayer/MultiscaleImageLayer.js#L88-L94
-  const correctedIndices = tileIndices.map((tile) => ({
-    x: tile.x,
-    y: tile.y,
-    z: Math.round(-tile.z + Math.log2(512 / tileSize)),
-  }));
-  return correctedIndices;
-}
-
-async function timeRegions({ file, sources, regions }) {
+async function timeRegions({ file, sources, regions }, iter) {
   for (const { url, format, tileSize, compression } of sources) {
     const loader = await getLoader({ url, format });
     const { height: rasterHeight, width: rasterWidth } = loader.getRasterSize({ z: 0 })
     const extent = [0, 0, rasterWidth, rasterHeight];
     for (const { id, yCoord, xCoord, zoom, viewports, numChannels } of regions) {
       for (const n of numChannels) {
-        const loaderSelection = [...Array(n).keys()].map(d => ({ channel: d }));
+        const loaderSelection = range(n).map(i => ({ channel: i }));
         for (const [height, width] of viewports) {
           const tileCoords = getTileCoords({
             yCoord,
@@ -65,6 +26,7 @@ async function timeRegions({ file, sources, regions }) {
           const records = times.map(([start, end], i) => {
             const { x, y, z } = tileCoords[i];
             return {
+              iter,
               file,
               format,
               compression,
@@ -93,7 +55,9 @@ async function timeRegions({ file, sources, regions }) {
 async function main() {
   console.time('Run benchmark.')
   for (const img of images) {
-    await timeRegions(img);
+    for (let i = 0; i < ITERS; i++) {
+      await timeRegions(img, i);
+    }
   }
   console.timeEnd('Run benchmark.')
 }
