--- src/test_config.json
@@ -1,25 +1,0 @@
-{
-  "images": [
-    {
-      "file": "original_file_name.img",
-      "sources": [
-        {
-          "url": "https://storage.googleapis.com/viv-demo/LuCa-7color_Scan1/",
-          "format": "zarr",
-          "compression": "blosc",
-          "tileSize": 1024
-        }
-      ],
-      "regions": [
-        { 
-          "id": "some-unique-id",
-          "xCoord": 0, 
-          "yCoord": 0, 
-          "zoom": -5,
-          "viewports": [[300, 400], [1000, 1000]],
-          "numChannels": [1, 2, 3, 4]
-        }
-      ]
-    }
-  ]
-}
